+++ components/ai.py
@@ -1,17 +1,19 @@
 import tcod
 
 class BasicMonster:
-    def take_turn(self, message, target, fov_map, game_map, entities):
+    def take_turn(self, target, fov_map, game_map, entities):
         """
-        message: 로그 띄우기
         target: 목표 대상
         fov_map, game_map, entities: 계산용
         """
+        results = []
         monster = self.owner
+        
         if fov_map.fov[monster.y, monster.x]:
-            if monster.distance_to(target) >= 2:
+            if monster.distance_to(target)  >= 2:
                 monster.move_astar(target, entities, game_map)
 
             elif target._Fighter.hp > 0:
-                monster._Fighter.attack(message, target)
-                #message.log(F'The {monster.name} insults you! Your ego is damaged!')
+                results.extend(monster._Fighter.attack(target))
+        
+        return results