+++ map_objects/game_map.py
@@ -13,6 +13,9 @@ from components.ai import BasicMonster
 from components.fighter import Fighter
 from components.luminary import Luminary
 
+# 렌더링
+from renderer.render_functions import RenderOrder
+
 class GameMap:
     def __init__(self, width, height):
         # 맵 크기 인자를 받아 객체의 높이와 너비 변수에 저장한다.
@@ -55,11 +58,13 @@ class GameMap:
             if randint(0, 100) < 80:
                 fighter_component = Fighter(hp=10, defense=0, power=3)
                 monster = Entity(nooks[i][1], nooks[i][0], '~', tcod.flame, 'crawling intestines',
-                                blocks=True, _Fighter=fighter_component, _Ai=ai_component)
+                                blocks=True, render_order=RenderOrder.ACTOR,
+                                _Fighter=fighter_component, _Ai=ai_component)
             else:
                 fighter_component = Fighter(hp=16, defense=1, power=4)
                 monster = Entity(nooks[i][1], nooks[i][0], 'S', tcod.dark_green, 'giant spider',
-                                blocks=True, _Fighter=fighter_component, _Ai=ai_component)
+                                blocks=True, render_order=RenderOrder.ACTOR,
+                                _Fighter=fighter_component, _Ai=ai_component)
             
             entities.append(monster)