+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   - Game map and cave are created with cellular automation method.
 - 카메라 구현
   - Camera implemented.
+- 간단한 전투와 사망 구현
+  - Simple combat and death implemented.
+- 간단한 메세지 로그
+  - Simple message log
 
 ## Changed
 - 타일셋 크기 32x32에서 16x16로 변경. 32x32도 남아는 있지만 한동안은 16x16 쓸 예정.